Update implementation for slf4j 2.0
slf4j 2.0 changes to use the ServiceLoader model to obtain the
implementation. So we must change. The bundle is no longer a fragment
and uses the OSGi Service Loader Mediator spec to provide the
implementation to the slf4j.api bundle.

Since we no longer need to implement a specific class, we refactor
into a new package.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>